Add custom strategy for toLazyByteString
This seems to halve the number of allocations when unstuffing. I left
this staged for 4 years, and I can't remember the original idea (besides
trying to allocate just once when converting). Tweaking the
smallChunkSize/firstsize doesn't seem to bring any improvement (even if
we know how many chunks the unstuffed string should be made of). Since
the tests pass and the weigh check show an improvement, I'm anyhow committing
this.